==> scripts/r/02_teams_reviews.Rmd <==
---
title: "02 - Teams reviews"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())
library(tidyverse)
library(magrittr)
library(glue)
```

```{r}
cols <- str_split("StartDate,EndDate,Status,IPAddress,Progress,Duration (in seconds),Finished,RecordedDate,ResponseId,RecipientLastName,RecipientFirstName,RecipientEmail,ExternalReference,LocationLatitude,LocationLongitude,DistributionChannel,UserLanguage,Q2,Q3,Q5_2,Q5_3,Q5_4,Q6,Q8,Q9,Q10,Q11,Q12,Q13,Q14,Q15", ",")

msa_reviews <- read_csv(list.files("data/peer_review/", pattern = ".csv", include.dirs = FALSE, full.names = TRUE), skip = 3, col_names = cols[[1]]) %>%
  mutate(
    q6 = case_when(
      Q6 == "deeply flawed and unpublishable." ~ 0,
      Q6 == "publishable with major revision." ~ 1,
      Q6 == "publishable with minor revision." ~ 2,
      Q6 == "publishable as is." ~ 3
    )
  ) %>%
  drop_na(Q3) %>%
  filter(Finished == TRUE)

saveRDS(msa_reviews, "./data/peer_review/msa_reviews.rds")
```

```{r analytic-quest}
dir.create("data/peer_review/msa_reviews/", showWarnings = FALSE)

for (team in unique(msa_reviews$Q3)) {
  file_name <- as.character(glue("data/peer_review/msa_reviews/{team}.csv"))
  
  msa_reviews %>%
    filter(Q3 == team) %>%
    select(-(StartDate:Q2)) %>%
    write_csv(file_name)
}
```
